Add cliseq/task/ref.py

+++ cliseq/cli/util.py
@@ -1,10 +1,11 @@
 #!/usr/bin/env python
 
+import bz2
+import gzip
 import logging
 import os
 import shutil
 from pathlib import Path
-from urllib.error import URLError
 from urllib.request import urlcleanup, urlopen, urlretrieve
 
 import yaml
@@ -22,25 +23,31 @@ def write_config_yml(path):
         )
 
 
-def download_hg(ver='hg38', work_dir='.'):
-    print_log('Download genome FASTA: {}'.format(ver))
+def fetch_executable(cmd):
+    executables = [
+        cp for cp
+        in [Path(p).joinpath(cmd) for p in ':'.split(os.environ['PATH'])]
+        if os.access(cp, os.X_OK)
+    ]
+    if executables:
+        return executables[0]
+    else:
+        raise RuntimeError(f'command not found: {cmd}')
+
+
+def download_hg(hg_ver='hg38', work_dir='.'):
+    print_log('Download genome FASTA: {}'.format(hg_ver))
     urls = read_yml(
         path=str(Path(__file__).parent.joinpath('../static/urls.yml'))
-    )['ref_fa_gz'][ver]
+    )['ref_fa_gz'][hg_ver]
     assert all([u.endswith('.gz') for u in urls]), 'invalid gzip URLs'
     output_path = Path(work_dir).joinpath(
-        '{}.fa.gz'.format('.'.join([Path(u).suffix.suffix.name for u in urls]))
+        '{}.fa.gz'.format('.'.join([Path(u).stem.stem.name for u in urls]))
     )
-    try:
-        if len(urls) == 1:
-            print_log(
-                'Retrieve:{0}{1} => {2}'.format(
-                    os.linesep, urls[0], output_path
-                )
-            )
-            urlretrieve(urls[0], filename=output_path)
-            urlcleanup()
-        else:
+    if len(urls) == 1:
+        retrieve_url(url=urls[0], output_path=output_path)
+    else:
+        try:
             with open(output_path, 'wb') as f:
                 for u in urls:
                     print_log(
@@ -49,13 +56,42 @@ def download_hg(ver='hg38', work_dir='.'):
                         )
                     )
                     f.write(urlopen(u).read())
-    except URLError as e:
-        logger = logging.getLogger(__name__)
-        logger.error(e)
-        os.remove(output_path)
+        except Exception as e:
+            if Path(output_path).exists():
+                os.remove(output_path)
+            raise e
+    print_log('Save: {}'.format(output_path))
+
+
+def retrieve_url(url, output_path):
+    try:
+        print_log(
+            'Retrieve:{0}{1} => {2}'.format(
+                os.linesep, url, output_path
+            )
+        )
+        urlretrieve(url, filename=output_path)
+        urlcleanup()
+    except Exception as e:
+        if Path(output_path).exists():
+            os.remove(output_path)
         raise e
+
+
+def open_readable_file(path):
+    if path.endswith('.gz'):
+        return gzip.open(path, 'rt')
+    elif path.endswith('.bz2'):
+        return bz2.open(path, 'rt')
     else:
-        print_log('Save: {}'.format(output_path))
+        return open(path, 'r')
+
+
+def remove_files_if_they_exists(*args):
+    for p in args:
+        if Path(p).exists():
+            print_log('Remove: {}'.format(p))
+            os.remove(p)
 
 
 def render_template(output_path, data=None, template=None):
